pdf_parse/pyproject.toml
[project]
name = "pdf-parse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "camelot-py>=1.0.9",
    "docling[vlm]>=2.66.0",
    "pdfplumber>=0.11.8",
    "rapidocr>=3.4.5",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

